+++ pkg/event/utils.go
