+++ pkg/redact/size.go
