+++ supergood_test.go
@@ -12,8 +12,8 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-	"github.com/supergoodsystems/supergood-go/internal/event"
-	remoteconfig "github.com/supergoodsystems/supergood-go/internal/remote-config"
+	"github.com/supergoodsystems/supergood-go/pkg/event"
+	remoteconfig "github.com/supergoodsystems/supergood-go/pkg/remote-config"
 )
 
 var events []*event.Event
